+++ ideas/src/app/core/definitions/Id.ts
@@ -1,0 +1,1 @@
+export type ID = string
