+++ ideas/src/app/core/definitions/MakerFunction.ts
@@ -1,0 +1,5 @@
+import { Maybe } from './Maybe'
+
+export type MakerFunction<Partial, Result> = (
+  input: Maybe<Partial, Result>
+) => Result
