+++ ideas/src/app/core/definitions/MakerFunctionFactory.ts
@@ -1,0 +1,5 @@
+import { MakerFunction } from './MakerFunction'
+
+export type MakerFunctionFactory<Deps, Partial, Result> = (
+  deps: Deps
+) => MakerFunction<Partial, Result>
