+++ ideas/src/app/core/definitions/Maybe.ts
@@ -1,0 +1,1 @@
+export type Maybe<T, S> = T | S
