+++ ideas/src/app/core/definitions/Serializeable.ts
@@ -1,0 +1,5 @@
+import { ID } from './Id'
+
+export type Serializable = {
+  id?: ID
+}
