+++ ideas/src/app/core/entities/entity.ts
@@ -1,0 +1,3 @@
+import { Serializable } from '../definitions/Serializeable'
+
+export interface Entity extends Serializable {}
